[CHG] code comments, renamed PostOverviewService to PostService, moved particle/particle.component to components/particle/particle.component, Lint Fixes
