+++ src/app/features/post-overview/components/post-dialog/post-dialog.component.ts
@@ -43,6 +43,7 @@ export class PostDialogComponent implements OnInit {
     }
     this.dialogRef.close(body);
   }
+
   onCancel(): void {
     this.dialogRef.close();
   }
