+++ src/app/models/particle.model.ts
